refactor: Split nav highlight script among multiple files to simplify logic
