+++ http/scripts/nav-highlights/constants.ts
@@ -1,0 +1,6 @@
+export interface Page {
+  pathFormat: RegExp;
+  navItemSelector: string;
+}
+
+export const ACTIVE_NAV_CLASS_NAME = 'nav-item-active';
