+++ http/templates/season_qualifiers.html.tera
@@ -2,6 +2,7 @@
 
 {% import "macros" as macros %}
 {% block scripts %}
+    <script type="module" src="/static/js/compiled/season-nav-highlights.js"></script>
     <script defer type="text/javascript" src="/static/js/season_qualifiers.js"></script>
 {% endblock %}
 {% block pagename %}{{season.format}} Qualifiers{% endblock %}
